Fix unknown id Type for EnumExt::deku_id()
Remove unwrap causing the following error when the EnumExt tried to infer
what type is was to return. An attribute could be added in the future to aid this,
but for now we will just not emit the deku_id() for this type of enum.
I decided in other causes (which aren't tested), to also remove the error
and just not emit the function also.

 help: message: called `Result::unwrap()` on an `Err` value: Error("expected `,`")

See #397