Closes #463: Add core_in_error support
Add the `core_in_error` feature and implement `core::error::Error`
for `DekuError` if the `core_in_error` feature is enabled.

This feature is automatically enabled in Rust nightly so when
using Rust nightly you get this feature by default. When using
stable Rust, this feature can be optionally added.