deku-derive: emit_enum(): Prefer repr over id_type for conversions
Use of the `id` attribute in enum contexts precludes use of
`id_type`, under which condition id_type is None. A None id_type
breaks the generated token stream at the point of the cast in the unsafe
expression accessing the discriminant value.

However, we must have a valid `repr` and that repr must match id_type
where it's present, by elimination of paths to the contrary in the
existing error handling. Emit the token for the repr type in place of
id_type for the purpose of the type conversion logic.

Signed-off-by: Andrew Jeffery <andrew@codeconstruct.com.au>